nc_scr: Add pmenu reference
Currently we use the fact that pmenu contains a struct nc_scr and use
the offset of the nc_scr member to convert from a struct nc_scr it's
pmenu.

To be able to do this conversion, nc_scr to pmenu, when the nc_scr
member is changed to a reference add a pmenu reference member to
struct nc_scr.

Signed-off-by: Geoff Levand <geoff@infradead.org>